// src/api/models/Crop.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Crop = {
  id: string;
  nameSv: string;
  nameEn: string;
  lifecycle: string;
  frostTolerance: string;
  minNightTempC: number | null;
  daylengthRequirement: string;
  notesSv: string | null;
  notesEn: string | null;
  createdAt: string;
};