// src/api/models/CropMethod.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type CropMethod = {
  id: string;
  cropId: string;
  labelSv: string;
  labelEn: string;
  germinationMinSoilTempC: number | null;
  germinationOptSoilTempC: number | null;
  daysToGerminationMin: number | null;
  daysToGerminationMax: number | null;
  daysToMaturityMin: number | null;
  daysToMaturityMax: number | null;
  transplantTolerance: string;
  gddRequired: number | null;
  plantBeforeFirstFrostDays: number | null;
  sortOrder: number;
  createdAt: string;
};